+++ Assets/NatShare/Plugins/Managed/Internal/PayloadBridge.cs
@@ -11,14 +11,15 @@ namespace NatShare.Internal {
     public static class PayloadBridge {
 
         private const string Assembly = @"__Internal";
+        public delegate void CompletionHandler (IntPtr context, bool success);
 
         #if UNITY_IOS && !UNITY_EDITOR
         [DllImport(Assembly, EntryPoint = @"NSCreateSharePayload")]
-        public static extern IntPtr CreateSharePayload (string subject, Action<IntPtr> completionHandler, IntPtr context);
+        public static extern IntPtr CreateSharePayload (string subject, CompletionHandler completionHandler, IntPtr context);
         [DllImport(Assembly, EntryPoint = @"NSCreateSavePayload")]
-        public static extern IntPtr CreateSavePayload (string album, Action<IntPtr> completionHandler, IntPtr context);
+        public static extern IntPtr CreateSavePayload (string album, CompletionHandler completionHandler, IntPtr context);
         [DllImport(Assembly, EntryPoint = @"NSCreatePrintPayload")]
-        public static extern IntPtr CreatePrintPayload (bool greyscale, bool landscape, Action<IntPtr> completionHandler, IntPtr context);
+        public static extern IntPtr CreatePrintPayload (bool greyscale, bool landscape, CompletionHandler completionHandler, IntPtr context);
         [DllImport(Assembly, EntryPoint = @"NSAddText")]
         public static extern void AddText (this IntPtr payload, string text);
         [DllImport(Assembly, EntryPoint = @"NSAddImage")]
@@ -29,9 +30,9 @@ namespace NatShare.Internal {
         public static extern void Commit (this IntPtr payload);
 
         #else
-        public static IntPtr CreateSharePayload (string subject, Action<IntPtr> completionHandler, IntPtr context) => IntPtr.Zero;
-        public static IntPtr CreateSavePayload (string album, Action<IntPtr> completionHandler, IntPtr context) => IntPtr.Zero;
-        public static IntPtr CreatePrintPayload (bool greyscale, bool landscape, Action<IntPtr> completionHandler, IntPtr context) => IntPtr.Zero;
+        public static IntPtr CreateSharePayload (string subject, CompletionHandler completionHandler, IntPtr context) => IntPtr.Zero;
+        public static IntPtr CreateSavePayload (string album, CompletionHandler completionHandler, IntPtr context) => IntPtr.Zero;
+        public static IntPtr CreatePrintPayload (bool greyscale, bool landscape, CompletionHandler completionHandler, IntPtr context) => IntPtr.Zero;
         public static void AddText (this IntPtr payload, string text) { }
         public static void AddImage (this IntPtr payload, IntPtr pixelBuffer, int width, int height) { }
         public static void AddMedia (this IntPtr payload, string uri) { }