+++ Assets/NatShare/Plugins/Managed/PrintPayload.cs
@@ -6,8 +6,7 @@
 namespace NatShare {
 
     using UnityEngine;
-    using System;
-    using System.Runtime.InteropServices;
+    using System.Threading.Tasks;
     using Internal;
 
     /// <summary>
@@ -22,33 +21,22 @@ namespace NatShare {
         /// </summary>
         /// <param name="greyscale">Optional. Should items be printed in greyscale</param>
         /// <param name="landscape">Optional. Should items be printed in landscape orientation</param>
-        /// <param name="completionHandler">Optional. Delegate invoked with whether saving was successful</param>
         [Doc(@"PrintPayloadCtor")]
-        public PrintPayload (bool greyscale = false, bool landscape = false, Action completionHandler = null) {
+        public PrintPayload (bool greyscale = false, bool landscape = false) {
+            this.commitTask = new TaskCompletionSource<bool>();
             switch (Application.platform) {
-                case RuntimePlatform.Android: {
-                    var nativePayload = new AndroidJavaObject(@"api.natsuite.natshare.PrintPayload", greyscale, landscape, PayloadAndroid.GetCallbackID(completionHandler));
-                    this.payload = new PayloadAndroid(nativePayload);
+                case RuntimePlatform.Android:
+                    this.payload = new PayloadAndroid(callback => new AndroidJavaObject(@"api.natsuite.natshare.PrintPayload", greyscale, landscape, callback));
                     break;
-                }
-                case RuntimePlatform.IPhonePlayer: {
-                    var callback = completionHandler != null ? (IntPtr)GCHandle.Alloc(completionHandler, GCHandleType.Normal) : IntPtr.Zero;
-                    var nativePayload = PayloadBridge.CreatePrintPayload(greyscale, landscape, PayloadiOS.OnCompletion, callback);
-                    this.payload = new PayloadiOS(nativePayload);
+                case RuntimePlatform.IPhonePlayer:
+                    this.payload = new PayloadiOS((callback, context) => PayloadBridge.CreatePrintPayload(greyscale, landscape, callback, context));
                     break;
-                }
                 default:
                     Debug.LogError("NatShare Error: PrintPayload is not supported on this platform");
                     this.payload = null; // Self-destruct >:D
                     break;
             }
         }
-
-        /// <summary>
-        /// Dispose the payload
-        /// </summary>
-        [Doc(@"Dispose")]
-        public void Dispose () => payload.Dispose();
 
         /// <summary>
         /// Nop. No concept as saving text to the gallery
@@ -68,8 +56,15 @@ namespace NatShare {
         /// <param name="path">Path to local media file to be shared</param>
         [Doc(@"AddMedia")]
         public void AddMedia (string uri) => payload.AddMedia(uri);
+
+        /// <summary>
+        /// Commit the payload and return whether payload was successfully shared
+        /// </summary>
+        [Doc(@"Commit")]
+        public Task<bool> Commit () => payload.Commit();
         #endregion
 
         private readonly ISharePayload payload;
+        private readonly TaskCompletionSource<bool> commitTask;
     }
 }