+++ iOS/NatShare/NSPrintPayload.m
@@ -50,7 +50,7 @@
     [printController presentAnimated:true completionHandler:^(UIPrintInteractionController* _, BOOL completed, NSError* error) {
         dispatch_async(dispatch_get_main_queue(), ^{
             if (completionHandler)
-                completionHandler();
+                completionHandler(completed);
         });
     }];
 }