+++ iOS/NatShare/NSSavePayload.m
@@ -73,7 +73,7 @@
     } completionHandler:^(BOOL success, NSError* error) {
         dispatch_async(dispatch_get_main_queue(), ^{
             if (completionHandler)
-                completionHandler();
+                completionHandler(success);
         });
     }];
 }